docbackend/requirements.txt
django==5.2
djangorestframework==3.16.0
django-cors-headers==4.3.1
pytesseract==0.3.13
pillow==11.2.1
transformers==4.37.2
python-docx==0.8.11
pypdf2==3.0.1
pdf2image==1.16.3
python-magic==0.4.27
psycopg2-binary==2.9.9
django-redis==5.4.0
redis==5.0.1
python-decouple==3.8
django-storages>=1.14.2
boto3>=1.34.0